gcf/requirements.txt
# Cloud Function dependencies
functions-framework==3.5.0
google-cloud-storage==2.14.0
google-cloud-bigquery==3.13.0
requests==2.31.0

